pyproject.toml
[project]
name = "tz-dimatech"
version = "0.1.0"
description = ""

readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "sqlalchemy (>=2.0.42,<3.0.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false